+++ src/internal/analysis/data_race_checker.hpp
@@ -4,16 +4,11 @@
 #include "coretrace_concurrency_analysis.hpp"
 #include "facts.hpp"
 
-namespace llvm
-{
-    class Module;
-} // namespace llvm
-
 namespace ctrace::concurrency::internal::analysis
 {
     class DataRaceChecker
     {
       public:
-        [[nodiscard]] DiagnosticReport run(const llvm::Module& module, const TUFacts& facts) const;
+        [[nodiscard]] DiagnosticReport run(const TUFacts& facts) const;
     };
 } // namespace ctrace::concurrency::internal::analysis